assets/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.5
Flask-Caching==2.0.2
Flask-Limiter==3.3.1
redis==4.6.0
sentry-sdk==1.29.2
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
mysqlclient==2.2.0
requests==2.31.0
Werkzeug==2.3.7
